requirements.txt
psycopg2-binary==2.9.9
pandas==2.2.2
python-dotenv==1.0.1
requests==2.32.3
tqdm==4.66.5
